Moved recently added unit tests to reply-error.test.js
